func_02057410: the digit loop divides unsigned, and the reloc says so
`bl` at 0x02057b64 relocated to `_ll_sdiv` (0x01ffaa34) where the ROM calls
`__aeabi_uldiv` (0x01ffa9dc). The bytes agreed anyway -- a relocated word is a
wildcard to the byte comparison -- so the function sat in the matched set as a
fakematch. `pr_linkcheck` is what sees it.

The sign is already consumed above the loop: the 0x80000000 test emits the minus
prefix and negates `v`, so every division below it is on a magnitude. Casting the
two operands to u64 selects the unsigned helper the ROM uses.

+++ src/func_02057410.c
@@ -299,9 +299,9 @@ int func_02057410(char *dst, int n, const char *fmt, int *args)
                         cnt = zero;
                         if (v != 0) {
                             do {
-                                int digit = (int)(v % pf[1]);
+                                int digit = (int)((u64)v % (u64)pf[1]);
                                 int dch;
-                                v = v / pf[1];
+                                v = (u64)v / (u64)pf[1];
                                 if (digit < 0xa) dch = digit + 0x30;
                                 else dch = digit + pf[2];
                                 numbuf[cnt] = dch;
